requirements.txt
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
flask-blueprint==1.3.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Pydantic==0.11.0
Flask-PyMongo==2.3.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==0.21.0
six==1.16.0
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
